reg, profile issues fixed

+++ src/app/admin/admin.component.ts
@@ -1,28 +1,34 @@
 ﻿import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { first } from 'rxjs/operators';
-
 import { User, Auth } from '@app/_models';
 import { AlertService, UserService, AuthenticationService } from '@app/_services';
 
 @Component({ templateUrl: 'admin.component.html' })
 export class AdminComponent implements OnInit, OnDestroy {
     currentUser: Auth;
+    userIsAdmin: boolean = false;
     currentUserSubscription: Subscription;
     users: User[] = [];
 
     constructor(
+        private router: Router,
         private authenticationService: AuthenticationService,
         private userService: UserService,
         private alertService: AlertService
     ) {
         this.currentUserSubscription = this.authenticationService.currentUser.subscribe(user => {
             this.currentUser = user;
+            if (this.currentUser) {
+              this.userIsAdmin = this.currentUser.user.userrole == 'admin';
+            }
         });
     }
 
     ngOnInit() {
-        this.loadAllUsers();
+        if (this.userIsAdmin) {this.loadAllUsers();}
+        else {this.router.navigate(['/']);}
     }
 
     ngOnDestroy() {
@@ -30,7 +36,8 @@ export class AdminComponent implements OnInit, OnDestroy {
         this.currentUserSubscription.unsubscribe();
     }
 
-    deleteUser(id: number) {
+    deleteUser(id: number, username: string) {
+      if (confirm(`Are you sure you want to delete the user '${username}'?'`)) {
         this.userService.delete(id).pipe(first()).subscribe(
           data => {
               this.alertService.success('Delete successful', true);
@@ -39,11 +46,11 @@ export class AdminComponent implements OnInit, OnDestroy {
           error => {
               this.alertService.error(error);
           });
+      }
     }
 
     private loadAllUsers() {
         this.userService.getAll().pipe(first()).subscribe(users => {
-            //console.log('admin.component.loadAllUsers | users:', users)
             this.users = users;
         }),
         (error => {
